requirements.txt
asn1crypto==1.0.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cloudinary==1.18.2
cryptography==2.7
Django==2.2.6
django-cloudinary-storage==0.2.3
django-cors-headers==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
mock==3.0.5
Pillow==6.2.0
psycopg2==2.8.3
pycparser==2.19
python-decouple==3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
